pyproject.toml
[project]
name = "data-extraction-vlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "azure-storage-blob>=12.25.1",
    "google-genai>=1.25.0",
    "google-generativeai>=0.8.5",
    "groq>=0.29.0",
    "instructor>=1.9.2",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "weave>=0.51.55",
]
